src/org/jessies/talc/AstSimplifier.java: add compile-time concatenation of string constants, and then rewrite the entire simplifier after realizing that it was throwing away information we'd added to AstNodes in earlier phases. The new implementation just modifies/replaces bits of the AST.
src/org/jessies/talc/AstNode.java: this means, unfortunately, that we need a whole load of new setters.

src/org/jessies/talc/Talc.java: wire up the new simplifier so it's possible for it to replace the topmost AstNode (or nodes), and add a debugging option to disable the simplification phase in case it's suspected of causing trouble.