src/org/jessies/talc/AstNode.java: ensure that a variable initialized with "null", if its type needs to be inferred, is inferred as "object".
src/org/jessies/talc/JvmCodeGenerator.java: treat the empty list and empty map as synonyms for list and map, since we implement generics using erasure, and anything that gets this far should have been type-checked and known to be correct.

src/org/jessies/talc/TalcType.java: fix the superclasses of the empty list and empty map to correspond to the superclasses of list and map. We'll have to be careful here when we start recording implemented interfaces, to make sure that these special types are kept in sync.

tests/compiler-trip.talc: test cases for the known bugs these fixes address.

TODO: job done.